demos/demo3-rag-poisoning/pyproject.toml
[project]
name = "demo3-rag-poisoning"
version = "0.1.0"
description = "Demo 3: RAG Poisoning - Data exfiltration via poisoned knowledge base documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flock-core",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
leak-server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
